+++ test/test_models.py
@@ -1,6 +1,6 @@
 import unittest
-from game.models import Tiles, BagTiles, Player, Board, Cell
-from game.scrabble_game import ScrabbleGame
+from game.models import Tiles, BagTiles, Player, Cell
+from game.board import Board
 from unittest.mock import patch
 
 class TestTiles(unittest.TestCase):
@@ -43,27 +43,11 @@ class TestBagTiles(unittest.TestCase):
             len(bag.tiles), 
             102,)
         
-
 class TestPlayer(unittest.TestCase):
     def test_init(self):
-        player_1 = Player()
-        self.assertEqual(len(player_1.tiles),0,)
-
-class TestBoard(unittest.TestCase):
-    def test_init(self):
-        board = Board()
-        self.assertEqual(len(board.grid),15,)
-        self.assertEqual(len(board.grid[0]),15,)
-
-class TestScrabbleGame(unittest.TestCase):
-    def test_init(self):
-        scrabble_game = ScrabbleGame(players_count=3)
-        self.assertIsNotNone(scrabble_game.board)
-        self.assertEqual(
-            len(scrabble_game.players),
-            3,
-        )
-        self.assertIsNotNone(scrabble_game.bag_tiles)
+        bag_tiles = BagTiles()
+        player_1 = Player(1, bag_tiles)
+        self.assertEqual(len(player_1.tiles),7,)
 
 class TestCell(unittest.TestCase):
     def test_init(self):
@@ -76,20 +60,18 @@ class TestCell(unittest.TestCase):
     def test_add_letter(self):
         cell = Cell(multiplier = 1, multiplier_type = 'letter')
         letter = Tiles(letter = 'P', value = 3)
-        cell.add_letter(letter = letter)
+        cell.add_letter(letter)
         self.assertEqual(cell.letter, letter)
 
     def test_cell_value(self):
-        cell = Cell(multiplier = 2, multiplier_type='letter')
+        board = Board()
+        board.positions()
+        cell= board.grid[11][0]
         letter = Tiles(letter = 'P', value = 3)
-        cell.add_letter(letter = letter)
+        cell.add_letter(letter)
         self.assertEqual(cell.calculate_value(),6)
 
-    def test_multiplier_word(self):
-        cell = Cell(multiplier = 2, multiplier_type= 'word')
-        letter = Tiles(letter = 'P', value = 3)
-        cell.add_letter(letter = letter)
-        self.assertEqual(cell.calculate_value(),3,)
+
 
 if __name__ == '__main__':
     unittest.main()